requirements.txt
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2.post0
enum34==1.1.6
flake8==3.5.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
